atc: behaviour: forbid negative limits
concourse/concourse#5421

Also lightly refactor the action validation logic so it happens a bit earlier,
at the time of unmarshalling the flag. The motivation for this was noticing
that there was logic to re-create the string form of the flag in order to
construct the error message, whereas in the `UnmarshalFlag` function the
unparsed string was still available.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>
Co-authored-by: Jamie Klassen <cklassen@pivotal.io>