atc/api: implement concurrent request limits
concourse/concourse#5421

We considered using semaphore from the golang.org/x/sync library, but
realized we don't need any waiting semantics so our implementation is a
bit simpler.

It is theoretically possible that the tests for pool that assert on
thread-safety could falsely pass on a sufficiently powerful/parallel
CPU, but these tests should be good enough for all practical purposes.

We briefly considered optimizing the `IsLimited` and `HandlerPool`
functions by replacing the `limits` field with a map rather than a list,
but the only performance impact seems to be at startup, and it might
require us to redesign the validation flow.

Signed-off-by: Bishoy Youssef <byoussef@pivotal.io>
Co-authored-by: Jamie Klassen <cklassen@pivotal.io>