atc: behaviour: return 501 when concurrent limit is 0
concourse/concourse#5471

In order to differentiate between an endpoint being temporarily
unavailable (i.e. the non-zero limit has been exceeded, but the endpoint
will eventually be available) and an endpoint that has been completely
disabled (by setting the concurrent request limit to 0), return a 501
error in the case of the latter. 501 is a "not implemented" error.

Use fakes in place of the real implementation of pools/policies to avoid
having concurrency in the tests, which could result in flakiness.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>